requirements.txt
deepstack_sdk==0.2.1
opencv_python==4.5.5.62
Pillow==9.0.1
streamlit==0.73.0
